+++ CHANGELOG.md
@@ -4,6 +4,7 @@ _The following changes are not released yet, but are on the master branch and wi
 
 **New features**
 
+* Added [`renderWeek`](http://react-day-picker.js.org/api/DayPicker`renderWeek) prop ([#497](https://github.com/gpbl/react-day-picker/issues/497) by [jenshandersson](https://github.com/jenshandersson))
 * Added [`onTodayButtonClick`](http://react-day-picker.js.org/api/DayPicker#onTodayButtonClick) prop ([#529](https://github.com/gpbl/react-day-picker/issues/529))
 
 ## [v6.2.0](https://github.com/gpbl/react-day-picker/tree/v6.2.0) (2017-10-05)